Kylista and the breastplate
Kylista can now be asked about the white breastplate even if flag 660 is set. The only downside here is if you know she owns it already, but then talk to her (use this code), and do not ask her about it, the flag is left unset, and this affects the equipment list scroll. I'd say that's a smaller bug than not being able to return it once 660 is set.