Marsten Evidence Scroll
If the incriminating scroll found in the Goblin King's treasure room that is signed by Marsten, Lord of Monitor, is read: 1) It sets the flag so you can accuse Marsten of some unspecified thing, and 2) Checks if Pomdirgun the Goblin King is dead, and if so sets the appropriate flag. See notes in the code for details.