SIFixes Adds Ernesto as NPC
Fixes for the following SI issues:
Ernesto duplicate spawning by egg.
Ernesto not checking for items before commenting.
Ernesto not correctly checking for the Banes being released, now
correctly comments on Shamino the MageLord.
Julia and Ernesto competing for beds in the Barracks with spawned
Rangers.
Topo now goes to the inn to eat.
Minor shape moves in the Blue Boar Inn, Shamino's Castle, and the
Ranger's Mess Hall.

Further details:
Map Edit in Shamino's Castle, a cobweb that overlays a nightstand at
1810,1012,1 has been moved south to make the nightstand accessible.
Map Edits in Moonshade, the chairs in the Rangers' mess hall have been
rearraigned so they can sit south of the table. The existing mislayered
plates and food have been changed, and smaller plates put in their
place, along with new but canonical blue wine glasses.
Fixes for hot bunking in the Rangers' barracks:
A nocturnal egg that creates a sleeping Ranger has removed from
2179,1844 so he does not compete for the bed Ernesto sleeps in.
Julia now sleeps in the tower at 2148,1814,6 to prevent her from using
Hawk's room if her original bed in the barracks is taken.
Julia has new items: a key to the tower (same as in the desk) and a
backpack to hold it in, and leather leggings.
Fixes the broken lunch schedule of Topo, who would get stuck in the
location of the cask rather than going to the inn to eat as intended.
Fixes the duplication of Ernesto by adding him as a single numbered NPC
and stopping the spawn eggs. His conversation has been improved
to look for items being discussed. He has a schedule and a magic bow to
make him stand out from the other rangers.