Add TouchUI_Android skeleton
iOS support uses a TouchUI class to manage several UI elements.  This patch adds a skeleton implementation for Android.  This is also the first Android-specific class added to the build, so this patch also introduces an android library that gets linked into Exult.  Any future android-specific code could be bundled into this library.