Add escape button
Without a keyboard, there is no way to hit the 'esc' button, but there are places in the game that this is required.  In particular, when looking through the telescope and in some error message screens.  Having an escape button is also helpful to clear all the open inventory pages and containers at once.  This patch adds a virtual escape button on the opposite side from the virtual joystick.