SIfixes add face for SmithzHorse, supporting files
Adds the graphics\faces folder for the face for SmithzHorse as created by Rhona. Number is padded to match my (Knight Captain's) mod.
Fixes the conversation for the nightmare horse shape to use that face, conversation is also split up to support different video resolutions. Note that the NPC has 0 health to start with if you test with a brand new game and hack-move.
Also adds one utility usecode in full, which could be used by other modders in original Exult games.