Sifixes: Fix possible hang in Fawn Trial cut scene.
Sometimes (not easily reproducible), the second part of the Fawn Trial
cut scene hangs after Lady Yelinda remarks that Kylista is not around.
This is because in those cases, she has the ASLEEP flag set and the
usecode she is supposed to run (saying "Zulith...") checks if somebody
can talk beforehand, which is not the case for sleeping NPCs. In order
to fix this, clear the ASLEEP flag for her in case it is set at the
beginning of the cutscene.

Fixes: #569